requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
groq==0.31.0
python-dotenv==1.0.1
langchain==0.1.11
langchain-groq==0.1.5
google-api-python-client==2.123.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0